+++ App.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import 'react-native-gesture-handler';
-import { Drawer } from './Navigator/Drawer';
+import { Drawer } from './src/navigation/Drawer';
 import { Provider } from 'react-redux';
-import { store } from './Redux/store';
+import { store } from './src/redux/store';
 
 export default function App() {
   return (
