+++ src/components/AnimatedListItem.tsx
